test: split tests in 2 suites
